chore: asset 관련 테이블 설계 및 ledger 테이블 수정함에 따라 entity 타입 일부 수정

+++ src/main/java/com/future/micklemuckle/modules/ledger/dto/CreateLedgerEntryRequest.java
@@ -3,6 +3,7 @@ package com.future.micklemuckle.modules.ledger.dto;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -21,7 +22,7 @@ public class CreateLedgerEntryRequest {
     private String entryType;
 
     @NotBlank
-    private int amount;
+    private BigDecimal amount;
 
     @NotBlank
     private String title;
